+++ pfnet/cconstr.pyx
@@ -10,6 +10,7 @@
 
 cimport cvec
 cimport cbus
+cimport cbus_dc
 cimport cconstr
 
 class ConstraintError(Exception):
@@ -686,13 +687,14 @@ cdef class CustomConstraint(ConstraintBase):
 
         pass
         
-    def count_step(self, bus, t):
+    def count_step(self, bus, busdc, t):
         """
         Performs count step.
 
         Parameters
         ----------
         bus : |Bus|
+        busdc : |BusDC|
         t : time period (int)
         """
         
@@ -712,25 +714,27 @@ cdef class CustomConstraint(ConstraintBase):
 
         pass
 
-    def analyze_step(self, bus, t):
+    def analyze_step(self, bus, busdc, t):
         """
         Performs analyze step.
        
         Parameters
         ----------
         bus : |Bus|
+        busdc : |BusDC|
         t : time period (int)
         """
         
         pass
 
-    def eval_step(self, bus, t, x, y=None):
+    def eval_step(self, bus, busdc, t, x, y=None):
         """
         Performs eval step.
        
         Parameters
         ----------
         bus : |Bus|
+        busdc : |BusDC|
         t : time period (int)
         x : |Array|
         y : |Array|
@@ -738,13 +742,14 @@ cdef class CustomConstraint(ConstraintBase):
  
         pass
 
-    def store_sens_step(self, bus, t, sA, sf, sGu, sGl):
+    def store_sens_step(self, bus, busdc, t, sA, sf, sGu, sGl):
         """
         Performs step for storing sensitivities.
        
         Parameters
         ----------
         bus : |Bus|
+        busdc : |BusDC|
         t : time period (int)
         sA : |Array|
         sf : |Array|
@@ -758,9 +763,9 @@ cdef void constr_init(cconstr.Constr* c):
     cdef CustomConstraint cc = <CustomConstraint>cconstr.CONSTR_get_data(c)
     cc.init()
 
-cdef void constr_count_step(cconstr.Constr* c, cbus.Bus* bus, int t):
+cdef void constr_count_step(cconstr.Constr* c, cbus.Bus* bus, cbus_dc.BusDC* busdc, int t):
     cdef CustomConstraint cc = <CustomConstraint>cconstr.CONSTR_get_data(c)
-    cc.count_step(new_Bus(bus),t)
+    cc.count_step(new_Bus(bus),new_BusDC(busdc),t)
 
 cdef void constr_allocate(cconstr.Constr* c):
     cdef CustomConstraint cc = <CustomConstraint>cconstr.CONSTR_get_data(c)
@@ -770,14 +775,14 @@ cdef void constr_clear(cconstr.Constr* c):
     cdef CustomConstraint cc = <CustomConstraint>cconstr.CONSTR_get_data(c)
     cc.clear()
 
-cdef void constr_analyze_step(cconstr.Constr* c, cbus.Bus* bus, int t):
+cdef void constr_analyze_step(cconstr.Constr* c, cbus.Bus* bus, cbus_dc.BusDC* busdc, int t):
     cdef CustomConstraint cc = <CustomConstraint>cconstr.CONSTR_get_data(c)
-    cc.analyze_step(new_Bus(bus),t)
+    cc.analyze_step(new_Bus(bus),new_BusDC(busdc),t)
 
-cdef void constr_eval_step(cconstr.Constr* c, cbus.Bus* bus, int t, cvec.Vec* v, cvec.Vec* ve):
+cdef void constr_eval_step(cconstr.Constr* c, cbus.Bus* bus, cbus_dc.BusDC* busdc, int t, cvec.Vec* v, cvec.Vec* ve):
     cdef CustomConstraint cc = <CustomConstraint>cconstr.CONSTR_get_data(c)
-    cc.eval_step(new_Bus(bus),t,Vector(v),Vector(ve))
+    cc.eval_step(new_Bus(bus),new_BusDC(busdc),t,Vector(v),Vector(ve))
 
-cdef void constr_store_sens_step(cconstr.Constr* c, cbus.Bus* bus, int t, cvec.Vec* sA, cvec.Vec* sf, cvec.Vec* sGu, cvec.Vec* sGl):
+cdef void constr_store_sens_step(cconstr.Constr* c, cbus.Bus* bus, cbus_dc.BusDC* busdc, int t, cvec.Vec* sA, cvec.Vec* sf, cvec.Vec* sGu, cvec.Vec* sGl):
     cdef CustomConstraint cc = <CustomConstraint>cconstr.CONSTR_get_data(c)
-    cc.eval_step(new_Bus(bus),t,Vector(sA),Vector(sf),Vector(sGu),Vector(sGl))
+    cc.eval_step(new_Bus(bus),new_BusDC(busdc),t,Vector(sA),Vector(sf),Vector(sGu),Vector(sGl))